--- src/Core/src/Hosting/MauiAppBuilderExtensions.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection.Extensions;
-
-namespace Microsoft.Maui.Hosting
-{
-	static class MauiAppBuilderExtensions
-	{
-		public static MauiAppBuilder ConfigureCoreServices(this MauiAppBuilder builder)
-		{
-			builder.Services.TryAddSingleton<IHashAlgorithm, Crc64>();
-			return builder;
-		}
-	}
-}
