--- src/Core/src/Services/IHashAlgorithm.cs
@@ -1,13 +1,0 @@
-﻿using System.IO;
-
-namespace Microsoft.Maui
-{
-	public interface IHashAlgorithm
-	{
-		byte[] ComputeHash(byte[] input);
-
-		byte[] ComputeHash(Stream inputStream);
-
-		string ComputeHashString(string input);
-	}
-}
